Check for and deal with unsuccessful network optimisations
PyPSA network optimisations can sometimes fail for various reasons.
Infeasibility is one reason, but that shouldn't happen for
well-designed experiments. However, anything reasons which can strike
sporadically even with feasible networks is numerical issues.

This commit adds code in the compute_optimum, mga and compute_near_opt
rules to check for and handle optimisation failures gracefully. In the
case of compute_optimum, a failure is considered critical and will now
cause the compute_optimum snakemake rule to fail, halting the entire
workflow. However, in the case of the mga and compute_near_opt rules,
we simply ignore the failed optimisation and continue. For the mga
rule it will just mean one fewer point, while the compute_near_opt
script will "re-try" the failed iteration with a new direction.