Use qhull directly to compute intersections
Instead of "probing" the intersection by optimising over it in many
directions, we use qhull from the scipy.spatial.HalfspaceIntersection
interface to compute the intersection. It still has to compute an
approximation (controlled by C, A, W options for qhull) but there is
now a better guarantee for result quality than before.

The max_iterations arguments has been removed from the `intersection`
function, while a `return_centre` argument has been added. This enable
the function to return the Chebyshev centre and radius of the
intersection, which are computed anyway since the
HalfspaceIntersection needs an interior point as initial input.