Select correct political map from NaturalEarth for README
The political map obtained from NaturalEarth by default attempts to
potrait state borders as they appear "de facto", not "de jure". They
thus consider Crimea as a part of Russia by default (see also
https://github.com/nvkelso/natural-earth-vector/issues/391), a choice
which accidentally made it into the map in our README in the form of
an illegitimate international border between Crimea and the rest of
Ukraine. The map is now fixed.