Fix qhull options in intersection computation
The intersection should be approximated instead of computed extactly
in order to avoid the generation of too many vertices. However, the
inclusion of the A0.99 and W0.01 options seems to have been too much;
the C option suffices. We increase the precision from C0.01 to C0.001.
This brings the number of vertices to around 6000 in our tests (down
from more than 100000 vertices when computed exactly).