Update pypsa-eur to version 0.6
The master branch of upstream pypsa-eur is merged into the local
pypsa-eur submodule. This comes with a few changes, most notably that
pypsa-eur now supposes named independent runs a la pypsa-eur-sec.

This makes our improvised support of this feature obsolete: we no
longer need to maintain separate pypsa-eur directories. In particular,
we can remove the "soft copy" function which copied the pypsa-eur base
directory, and the associated logic in the top-level Snakefile.

The pypsa-eur sections of the config files were also updated to
reflect upstream updates in how carriers are specified.